pyproject.toml
[tool.poetry]
name = "rmqr-decoder"
version = "0.1.0"
description = ""
authors = ["organic-nailer <hykwyuk0125@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytesseract = "^0.3.10"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
